Create a bootstrapped tibble

bootstrapify() creates a bootstrapped tibble with virtual groups.

bootstrapify(data, times, ..., key = ".bootstrap")

Arguments

data
A tbl.

times
A single integer specifying the number of resamples. If the tibble is grouped, this is the number of resamples per group.

...
Not used.

key
A single character specifying the name of the virtual group that is added.

Value

A resampled_df with an extra group specified by key.

Details

Currently you can use dplyr::summarise(), dplyr::do(), and the group manipulation functions: dplyr::group_map(), dplyr::group_walk(), dplyr::group_nest() and dplyr::group_split() on a resampled_df.

See also

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(broom)

bootstrapify(iris, 5)

#> # A tibble: 150 x 5
#> # Groups:   .bootstrap [5]
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>  
#>  1          5.1         3.5          1.4         0.2 setosa 
#>  2          4.9         3            1.4         0.2 setosa 
#>  3          4.7         3.2          1.3         0.2 setosa 
#>  4          4.6         3.1          1.5         0.2 setosa 
#>  5          5           3.6          1.4         0.2 setosa 
#>  6          5.4         3.9          1.7         0.4 setosa 
#>  7          4.6         3.4          1.4         0.3 setosa 
#>  8          5           3.4          1.5         0.2 setosa 
#>  9          4.4         2.9          1.4         0.2 setosa 
#> 10          4.9         3.1          1.5         0.1 setosa 
#> # … with 140 more rows

iris %>%
  bootstrapify(5) %>%
  summarise(per_strap_mean = mean(Petal.Width))

#> # A tibble: 5 x 2
#>   .bootstrap per_strap_mean
#>        <int>          <dbl>
#> 1          1           1.12
#> 2          2           1.23
#> 3          3           1.18
#> 4          4           1.10
#> 5          5           1.14

iris %>%
  group_by(Species) %>%
  bootstrapify(5) %>%
  summarise(per_strap_species_mean = mean(Petal.Width))

#> # A tibble: 15 x 3
#> # Groups:   Species [3]
#>    Species    .bootstrap per_strap_species_mean
#>    <fct>           <int>                  <dbl>
#>  1 setosa              1                  0.232
#>  2 setosa              2                  0.264
#>  3 setosa              3                  0.246
#>  4 setosa              4                  0.248
#>  5 setosa              5                  0.244
#>  6 versicolor          1                  1.31 
#>  7 versicolor          2                  1.29 
#>  8 versicolor          3                  1.33 
#>  9 versicolor          4                  1.30 
#> 10 versicolor          5                  1.33 
#> 11 virginica           1                  2.05 
#> 12 virginica           2                  1.99 
#> 13 virginica           3                  2.05 
#> 14 virginica           4                  1.99 
#> 15 virginica           5                  2.02 

iris %>%
  bootstrapify(5) %>%
  do(tidy(lm(Sepal.Width ~ Sepal.Length + Species, data = .)))

#> # A tibble: 20 x 6
#> # Groups:   .bootstrap [5]
#>    .bootstrap term              estimate std.error statistic  p.value
#>         <int> <chr>                <dbl>     <dbl>     <dbl>    <dbl>
#>  1          1 (Intercept)          1.54     0.281       5.49 1.74e- 7
#>  2          1 Sepal.Length         0.373    0.0560      6.66 5.27e-10
#>  3          1 Speciesversicolor   -1.00     0.0824    -12.2  6.35e-24
#>  4          1 Speciesvirginica    -1.09     0.112      -9.73 1.45e-17
#>  5          2 (Intercept)          1.34     0.256       5.23 5.69e- 7
#>  6          2 Sepal.Length         0.416    0.0498      8.36 4.63e-14
#>  7          2 Speciesversicolor   -1.05     0.0691    -15.2  6.57e-32
#>  8          2 Speciesvirginica    -1.07     0.0836    -12.8  1.14e-25
#>  9          3 (Intercept)          1.86     0.222       8.38 4.10e-14
#> 10          3 Sepal.Length         0.312    0.0430      7.26 2.08e-11
#> 11          3 Speciesversicolor   -0.885    0.0714    -12.4  1.39e-24
#> 12          3 Speciesvirginica    -0.990    0.0827    -12.0  1.98e-23
#> 13          4 (Intercept)          1.51     0.268       5.63 8.81e- 8
#> 14          4 Sepal.Length         0.377    0.0524      7.18 3.23e-11
#> 15          4 Speciesversicolor   -1.01     0.0742    -13.6  9.49e-28
#> 16          4 Speciesvirginica    -0.920    0.0952     -9.66 2.19e-17
#> 17          5 (Intercept)          1.78     0.215       8.28 7.32e-14
#> 18          5 Sepal.Length         0.347    0.0420      8.27 7.41e-14
#> 19          5 Speciesversicolor   -1.07     0.0745    -14.3  1.37e-29
#> 20          5 Speciesvirginica    -1.15     0.0930    -12.4  1.41e-24

# Alternatively, use the newer group_map()
iris %>%
  bootstrapify(5) %>%
  group_map(~tidy(lm(Sepal.Width ~ Sepal.Length + Species, data = .x)))

#> # A tibble: 20 x 6
#> # Groups:   .bootstrap [5]
#>    .bootstrap term              estimate std.error statistic  p.value
#>         <int> <chr>                <dbl>     <dbl>     <dbl>    <dbl>
#>  1          1 (Intercept)          1.44     0.224       6.43 1.74e- 9
#>  2          1 Sepal.Length         0.406    0.0446      9.12 5.54e-16
#>  3          1 Speciesversicolor   -1.07     0.0678    -15.7  2.93e-33
#>  4          1 Speciesvirginica    -1.17     0.0882    -13.3  5.86e-27
#>  5          2 (Intercept)          1.77     0.205       8.62 1.02e-14
#>  6          2 Sepal.Length         0.321    0.0402      7.98 3.86e-13
#>  7          2 Speciesversicolor   -0.831    0.0666    -12.5  8.57e-25
#>  8          2 Speciesvirginica    -0.888    0.0900     -9.87 6.39e-18
#>  9          3 (Intercept)          1.47     0.245       6.00 1.50e- 8
#> 10          3 Sepal.Length         0.391    0.0485      8.05 2.63e-13
#> 11          3 Speciesversicolor   -1.01     0.0734    -13.8  3.92e-28
#> 12          3 Speciesvirginica    -1.13     0.0979    -11.5  2.76e-22
#> 13          4 (Intercept)          1.52     0.225       6.74 3.40e-10
#> 14          4 Sepal.Length         0.379    0.0445      8.52 1.86e-14
#> 15          4 Speciesversicolor   -0.982    0.0704    -14.0  1.19e-28
#> 16          4 Speciesvirginica    -0.978    0.0925    -10.6  9.70e-20
#> 17          5 (Intercept)          1.96     0.226       8.66 7.95e-15
#> 18          5 Sepal.Length         0.279    0.0454      6.15 7.11e- 9
#> 19          5 Speciesversicolor   -0.928    0.0763    -12.2  6.09e-24
#> 20          5 Speciesvirginica    -0.814    0.103      -7.93 5.13e-13

# Alter the name of the group with `key`
# Materialize them with collect()
straps <- bootstrapify(iris, 5, key = ".straps")
collect(straps)

#> # A tibble: 750 x 6
#> # Groups:   .straps [5]
#>    .straps Sepal.Length Sepal.Width Petal.Length Petal.Width Species   
#>      <int>        <dbl>       <dbl>        <dbl>       <dbl> <fct>     
#>  1       1          5.1         2.5          3           1.1 versicolor
#>  2       1          4.8         3.4          1.9         0.2 setosa    
#>  3       1          6.5         3            5.8         2.2 virginica 
#>  4       1          4.8         3.4          1.6         0.2 setosa    
#>  5       1          5.8         2.8          5.1         2.4 virginica 
#>  6       1          5.6         2.5          3.9         1.1 versicolor
#>  7       1          6.1         2.8          4           1.3 versicolor
#>  8       1          6.6         3            4.4         1.4 versicolor
#>  9       1          6.2         3.4          5.4         2.3 virginica 
#> 10       1          5.5         2.4          3.7         1   versicolor
#> # … with 740 more rows