Force virtual groups to become explicit rows

When collect() is used on a resampled_df, the virtual bootstrap groups are made explicit.

# S3 method for resampled_df
collect(x, ..., id = NULL, original_id = NULL)

Arguments

x
A resampled_df.

...
Not used.

id
Optional. A single character that specifies a name for a column containing a sequence from 1:n for each bootstrap group.

original_id
Optional. A single character that specifies a name for a column containing the original position of the bootstrapped row.

Examples

library(dplyr)

# virtual groups become real rows
collect(bootstrapify(iris, 5))

#> # A tibble: 750 x 6
#> # Groups:   .bootstrap [5]
#>    .bootstrap Sepal.Length Sepal.Width Petal.Length Petal.Width Species   
#>         <int>        <dbl>       <dbl>        <dbl>       <dbl> <fct>     
#>  1          1          6.2         3.4          5.4         2.3 virginica 
#>  2          1          6.1         2.6          5.6         1.4 virginica 
#>  3          1          4.6         3.6          1           0.2 setosa    
#>  4          1          5.8         2.7          4.1         1   versicolor
#>  5          1          5.8         2.8          5.1         2.4 virginica 
#>  6          1          5.4         3            4.5         1.5 versicolor
#>  7          1          6.3         2.8          5.1         1.5 virginica 
#>  8          1          6.7         3            5.2         2.3 virginica 
#>  9          1          5.6         2.7          4.2         1.3 versicolor
#> 10          1          6.5         3.2          5.1         2   virginica 
#> # … with 740 more rows

# add on the id column for an identifier per bootstrap
collect(bootstrapify(iris, 5), id = ".id")

#> # A tibble: 750 x 7
#> # Groups:   .bootstrap [5]
#>    .bootstrap   .id Sepal.Length Sepal.Width Petal.Length Petal.Width Species   
#>         <int> <int>        <dbl>       <dbl>        <dbl>       <dbl> <fct>     
#>  1          1     1          6.9         3.2          5.7         2.3 virginica 
#>  2          1     2          6.7         3.3          5.7         2.5 virginica 
#>  3          1     3          6.1         2.9          4.7         1.4 versicolor
#>  4          1     4          5.6         2.7          4.2         1.3 versicolor
#>  5          1     5          6.3         2.5          4.9         1.5 versicolor
#>  6          1     6          6.4         3.2          5.3         2.3 virginica 
#>  7          1     7          5.1         3.8          1.9         0.4 setosa    
#>  8          1     8          6.7         3.3          5.7         2.5 virginica 
#>  9          1     9          4.5         2.3          1.3         0.3 setosa    
#> 10          1    10          4.4         2.9          1.4         0.2 setosa    
#> # … with 740 more rows

# add on the original_id column to know which row this bootstrapped row
# originally came from
collect(bootstrapify(iris, 5), original_id = ".original_id")

#> # A tibble: 750 x 7
#> # Groups:   .bootstrap [5]
#>    .bootstrap .original_id Sepal.Length Sepal.Width Petal.Length Petal.Width
#>         <int>        <int>        <dbl>       <dbl>        <dbl>       <dbl>
#>  1          1          131          7.4         2.8          6.1         1.9
#>  2          1          139          6           3            4.8         1.8
#>  3          1           99          5.1         2.5          3           1.1
#>  4          1           32          5.4         3.4          1.5         0.4
#>  5          1           25          4.8         3.4          1.9         0.2
#>  6          1           72          6.1         2.8          4           1.3
#>  7          1           47          5.1         3.8          1.6         0.2
#>  8          1           28          5.2         3.5          1.5         0.2
#>  9          1          126          7.2         3.2          6           1.8
#> 10          1           52          6.4         3.2          4.5         1.5
#> # … with 740 more rows, and 1 more variable: Species <fct>